requirements.txt
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
defusedxml==0.7.1
dj-database-url==3.0.0
Django==4.2.21
django-cors-headers==4.7.0
django-rest-passwordreset==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
gunicorn==23.0.0
h11==0.16.0
httpie==3.2.4
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.4
packaging==25.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
whitenoise==6.9.0
